# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Telegram Bot API - Focus Sessions Application
  description: |
    API для мини-приложения синхронных сессий фокуса с интеграцией Telegram Bot API.
    
    ## Аутентификация
    API использует HTTP-only cookies для аутентификации (защита от XSS атак).
    Токены устанавливаются автоматически при входе через endpoint `/api/v1/auth/login`.
    Access token доступен для всех API endpoints, refresh token - только для `/auth/refresh`.
    
    ## Особенности
    - Solo и Group сессии фокуса
    - Real-time обновления через WebSocket
    - Интеграция с Telegram Bot API
    - Система лидерборда
    - Управление задачами
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Локальный сервер разработки
  - url: https://api.example.com/api/v1
    description: Production сервер
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT токен в HTTP-only cookie access_token (устанавливается автоматически при /auth/login).
        Для обратной совместимости также поддерживается Bearer токен в Authorization header.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Сообщение об ошибке
        code:
          type: string
          description: Код ошибки
        message:
          type: string
          description: Детальное описание ошибки
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
...
    AuthRequest:
      type: object
      properties:
        initData:
          type: string
          description: |
            URL-encoded строка инициализации Telegram Mini App (query_id, user, auth_date, hash).
            Должна передаваться в точном виде, как получено от Telegram WebApp.
          example: "query_id=AAH123&user=%7B%22id%22%3A123%2C%22first_name%22%3A%22Ivan%22%7D&auth_date=1712147200&hash=b1a2c3"
        deviceId:
          type: string
          description: |
            Идентификатор устройства/клиента (например, UUID браузера или мобильного устройства).
            Используется для трекинга сессий и отзывов refresh токенов.
          example: "device-f5e3b94a-5e1b-4a16-9a3f-0db8741c4f19"
      required:
        - initData
        - deviceId
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
      required:
        - user
      description: |
        Токены устанавливаются в HTTP-only cookies:
        - access_token: доступен для всех API endpoints (Path=/)
        - refresh_token: доступен только для /auth/refresh (Path=/api/v1/auth/refresh)
    RefreshTokenResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
      description: |
        Новый access_token устанавливается в HTTP-only cookie.
        Refresh token читается из cookie refresh_token.
...
paths:
  /auth/login:
    post:
      tags:
        - auth
      summary: Вход через Telegram initData
      description: |
        Авторизация пользователя через Telegram Mini App `initData`.
        `initData` содержит URL-encoded параметры (query_id, user, auth_date, hash), полученные на фронтенде из Telegram WebApp.
        Токены устанавливаются в HTTP-only cookies для защиты от XSS атак.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Успешная авторизация
          headers:
            Set-Cookie:
              description: |
                HTTP-only cookies с токенами:
                - access_token: HttpOnly; Secure; SameSite=Strict; Path=/; Max-Age=900
                - refresh_token: HttpOnly; Secure; SameSite=Strict; Path=/api/v1/auth/refresh; Max-Age=604800
              schema:
                type: string
                example: access_token=eyJhbGc...; HttpOnly; Secure; SameSite=Strict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Ошибка авторизации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags:
        - auth
      summary: Обновление access token
      description: |
        Получение нового access token используя refresh token из cookie.
        Refresh token читается из HTTP-only cookie refresh_token.
      responses:
        '200':
          description: Токен обновлён
          headers:
            Set-Cookie:
              description: |
                Новый access_token в HTTP-only cookie:
                - access_token: HttpOnly; Secure; SameSite=Strict; Path=/; Max-Age=900
              schema:
                type: string
                example: access_token=eyJhbGc...; HttpOnly; Secure; SameSite=Strict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '401':
          description: Невалидный refresh token или отсутствует cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags:
        - auth
      summary: Выход из системы
      description: Выход пользователя из системы и инвалидация токенов
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Успешный выход
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...